app/requirements.txt
streamlit
pandas
numpy
langchain
sentence_transformers
pinecone-client